Add files via upload
This portfolio project covered data tidying and cleaning. This module covered, though wasn't limited to:
- Combining multiple csvs into one pandas dataframe
- Regular expressions
- Splitting data by characters
- Removing duplicates
- Dealing with empty values
All of these principles are applied to state census data, here present in nine different CSV files. Data are cleaned and sorted (steps are clearly labelled with comments) so that basic calculations can be performed and visualized.